test: step4 테스트 패키지 클래스 이동, LineTest -> LadderLineTest
