+++ src/test/java/nextstep/ladder/mock/MockAlwaysTruePointsGenerator.java
@@ -1,0 +1,17 @@
+package nextstep.ladder.mock;
+
+import nextstep.ladder.util.pointsgenerator.PointsGenerator;
+
+import java.util.List;
+
+public class MockAlwaysTruePointsGenerator implements PointsGenerator {
+    @Override
+    public List<Boolean> getPoints(int pointsSize) {
+        return null;
+    }
+
+    @Override
+    public boolean generatePoint() {
+        return true;
+    }
+}
